Revert "#36 Restaurant detail modal"
