+++ Frontend_WhatToWeEat/src/components/RestaurantPick.vue
@@ -18,18 +18,19 @@
         class="w-full flex flex-col sm:flex-row"
       >
         <!--給圖片的畫框-->
-        <div class="w-full h-[50vw] sm:w-64 sm:h-36 bg-wte-bg">
-          <img
+        <div class="w-full h-[50vw] sm:w-64 sm:h-36 bg-wte-warning">
+          <n-image
             v-if="recommendedRest.value.image"
-            class="h-full w-full object-cover"
+            object-fit="cover"
             :src="recommendedRest.value.image"
-          >
-          <img
+            lazy="true"
+          />
+          <n-image
             v-else
-            class="h-full w-full object-cover"
+            object-fit="cover"
             :src="fakeBannerImg"
             lazy="true"
-          >
+          />
         </div>
 
         <div class="p-4 flex flex-col justify-center">
@@ -58,6 +59,14 @@
         <i-material-symbols:add />
         新增餐廳
       </n-button>
+      <n-button
+        type="primary"
+        class="bg-wte-primary"
+      >
+        <i-material-symbols:edit />
+
+        編輯餐廳
+      </n-button>
     </div>
 
     <!-- 手機板功能鍵區 -->
@@ -82,16 +91,127 @@
       >
         <i-material-symbols:add />
       </n-button>
+      <n-button
+        circle
+        size="large"
+        type="primary"
+        class="bg-wte-primary"
+      >
+        <i-material-symbols:edit />
+      </n-button>
     </div>
 
     <!-- 新增餐廳 modal -->
-    <RestaurantModal
-      mode="create"
-      :is-show-rest-modal="isShowAddRestModal"
-      :is-show-loading="isShowFormLoading"
-      @finish="postRestaurant"
-      @close="isShowAddRestModal = false"
-    />
+    <n-modal
+      v-model:show="isShowAddRestModal"
+      :on-after-leave="resetForm"
+    >
+      <n-spin
+        :show="isShowFormLoading"
+        class="bg-white"
+      >
+        <template #description>
+          稍等我上傳個圖片
+        </template>
+        <n-card
+          style="width: 600px"
+          title="新增餐廳"
+          :bordered="false"
+          size="huge"
+          role="dialog"
+          aria-modal="true"
+        >
+          <n-form
+            ref="formRef"
+            :model="restaurantForm"
+            :rules="restFormRules"
+          >
+            <n-grid
+              x-gap="12"
+              y-gap="12"
+              cols="2"
+              responsive="screen"
+            >
+              <n-form-item-gi
+                label="餐廳名稱"
+                path="name"
+              >
+                <n-input
+                  v-model:value="restaurantForm.name"
+                  placeholder="餐廳名稱"
+                />
+              </n-form-item-gi>
+              <n-form-item-gi
+                label="餐廳電話"
+                path="phone"
+              >
+                <n-input
+                  v-model:value="restaurantForm.phone"
+                  placeholder="0912345678"
+                />
+              </n-form-item-gi>
+              <n-form-item-gi
+                label="餐廳地址"
+                path="address"
+              >
+                <n-input
+                  v-model:value="restaurantForm.address"
+                  placeholder="餐廳地址"
+                />
+              </n-form-item-gi>
+            </n-grid>
+          </n-form>
+
+          <p class="mb-1">
+            圖片上傳
+          </p>
+          <n-upload
+            ref="imgUploadRef"
+            directory-dnd
+            accept=".jpg, .jpeg, .png"
+            :max="1"
+            list-type="image"
+            :on-before-upload="handleFileUpload"
+          >
+            <n-upload-dragger>
+              <div class="flex items-center justify-center w-full h-14">
+                <i-material-symbols:cloud-upload class="h-10 w-10" />
+              </div>
+              <n-text style="font-size: 16px">
+                點擊或拖曳來上傳圖片
+              </n-text>
+              <n-p
+                depth="3"
+                style="margin: 8px 0 0 0"
+              >
+                檔案格式限制為JPG, JPEG, PNG且大小不超過3MB
+              </n-p>
+            </n-upload-dragger>
+          </n-upload>
+
+          <template #footer>
+            <div class="flex w-full gap-3">
+              <n-button
+                type=""
+                size="large"
+                class="ml-auto"
+                @click="isShowAddRestModal = false"
+              >
+                取消
+              </n-button>
+              <n-button
+                type="primary"
+                size="large"
+                class="bg-wte-primary"
+                @click="handleInputValidation"
+              >
+                送出
+              </n-button>
+            </div>
+          </template>
+        </n-card>
+      </n-spin>
+    </n-modal>
   </div>
 </template>
 
@@ -107,7 +227,69 @@ const store = useStore();
 const isRecommend = ref(false); // isRecommend : false:按鈕狀態 true顯示餐廳資訊
 const recommendedRest = reactive({ value: [] });
 const isShowAddRestModal = ref(false);
+const restaurantForm = reactive({
+  name: "",
+  phone: "",
+  address: "",
+  image: "",
+});
+const restFormRules = ref({
+  name: {
+    required: true,
+    message: "請輸入餐廳名稱",
+    trigger: ["input", "blur"],
+  },
+  phone: {
+    required: true,
+    validator(rule, value) {
+      if (!value) {
+        return new Error("請輸入電話");
+      } else if (!/0\d{9}/.test(value)) {
+        return new Error("電話格式錯誤");
+      }
+      return true;
+    },
+    trigger: ["input", "blur"],
+  },
+  address: {
+    required: true,
+    message: "請輸入餐廳地址",
+    trigger: ["input", "blur"],
+  },
+});
 const isShowFormLoading = ref(false);
+const uploadImg = ref(new FormData());
+const formRef = ref(null);
+const imgUploadRef = ref(null);
+
+function resetForm() {
+  Object.keys(restaurantForm).forEach((key) => {
+    restaurantForm[key] = "";
+  });
+  imgUploadRef.value.clear();
+}
+
+function handleFileUpload(options) {
+  let format = options.file.file.type;
+  let sizeMb = options.file.file.size / 1024 / 1024;
+
+  if (
+    format != "image/png" &&
+    format != "image/jpg" &&
+    format != "image/jpeg"
+  ) {
+    message.warning("圖片格式錯誤，僅支持JPG, JPEG, PNG");
+    imgUploadRef.value.clear();
+    return false;
+  }
+
+  if (sizeMb > 3) {
+    message.warning("圖片大小過大，請保持在3MB以下");
+    imgUploadRef.value.clear();
+    return false;
+  }
+  uploadImg.value.append("restImg", options.file.file);
+}
 
 function getRecommendButton() {
   getRestaurantPick();
@@ -123,22 +305,31 @@ const getRestaurantPick = async () => {
   }
 };
 
-const postRestaurant = async (data) => {
+const postRestaurant = async () => {
   try {
     isShowFormLoading.value = true;
-    const res = await apiPostRestaurant(data.restaurantForm);
-    if (data.img.has("restImg")) {
-      await apiPostRestImg(data.img, res.data.id);
-    }
+    const res = await apiPostRestaurant(restaurantForm);
+    const imgRes = await apiPostRestImg(uploadImg.value, res.data.id);
     store.getRestaurantData();
+    isShowAddRestModal.value = false;
+    isShowFormLoading.value = false;
+    resetForm();
     message.success("餐廳新增成功");
   } catch (err) {
+    isShowAddRestModal.value = false;
+    isShowFormLoading.value = false;
     message.error("餐廳新增失敗");
     console.error(err);
   }
+};
 
-  isShowAddRestModal.value = false;
-  isShowFormLoading.value = false;
+const handleInputValidation = async (e) => {
+  e.preventDefault();
+  formRef.value?.validate((errors) => {
+    if (!errors) {
+      postRestaurant();
+    }
+  });
 };
 </script>
